+++ src/components/modals/ModalBackdrop.css
@@ -1,0 +1,6 @@
+#modal-backdrop {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-color: var(--black-opacity-50);
+}
