+++ src/components/modals/nutritionModal/NutritionDetailsModal.css
@@ -1,4 +1,8 @@
 #nutri-label {
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
     border: 2px solid black;
     width: 270px;
     padding: 0 7px;
