+++ src/styles/variables.css
@@ -1,4 +1,8 @@
 :root {
+    /* SPACING ----------------------------------------------------------------------------------- */
+    --app-height-100: calc(100% - 60px);
+
+    /* COLORS ------------------------------------------------------------------------------------ */
     /* Base Colors */
     --black: #161616;
     --white: #fff;
